feat: update user and token models to use int64 for IDs, add base repository interface, and implement user repository methods

+++ internal/domain/repository/user_repository.go
@@ -9,11 +9,11 @@ import (
 
 type UserRepository interface {
 	Create(ctx context.Context, user *entity.User) error
-	FindByID(ctx context.Context, id uint) (*entity.User, error)
+	FindByID(ctx context.Context, id int64) (*entity.User, error)
 	FindByUsername(ctx context.Context, username string) (*entity.User, error)
 	FindByEmail(ctx context.Context, email string) (*entity.User, error)
 	Update(ctx context.Context, user *entity.User) error
-	SoftDelete(ctx context.Context, id uint) error
+	SoftDelete(ctx context.Context, id int64) error
 }
 
 var (
